Backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
boto3==1.28.85
Pillow==10.0.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
